Moved ProductItemDiffCallback to it's own file
