+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/ProductListAdapter.kt
@@ -4,7 +4,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.selection.SelectionTracker
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.woocommerce.android.analytics.AnalyticsEvent.PRODUCT_LIST_PRODUCT_TAPPED
 import com.woocommerce.android.analytics.AnalyticsTracker
@@ -48,22 +47,6 @@ class ProductListAdapter(
 
         if (position == itemCount - 1) {
             loadMoreListener.onRequestLoadMore()
-        }
-    }
-
-    object ProductItemDiffCallback : DiffUtil.ItemCallback<Product>() {
-        override fun areItemsTheSame(
-            oldItem: Product,
-            newItem: Product
-        ): Boolean {
-            return oldItem.remoteId == newItem.remoteId
-        }
-
-        override fun areContentsTheSame(
-            oldItem: Product,
-            newItem: Product
-        ): Boolean {
-            return oldItem == newItem
         }
     }
 }
